Skylark-AI_Agent/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
groq==1.0.0
pydantic==2.6.3
python-multipart==0.0.9
httpx==0.27.0
tenacity==8.2.3
